+++ src/entrypoints/devtools/main.ts
@@ -1,5 +1,5 @@
 import { browser } from 'wxt/browser'
 
-const panelTitle = `${import.meta.env.FIREFOX ? '' : '☄️'} Meteor`
+const panelTitle = 'Meteor'
 
 browser.devtools.panels.create(panelTitle, '', '/devtools-panel.html')